refactor: split the selector module
